+++ public/assignment/views/website/controllers/website-edit.controller.client.js
@@ -24,24 +24,32 @@
         vm.updateWebsite = updateWebsite;
 
         function deleteWebsite() {
-            var promise = WebsiteService.deleteWebsite(vm.websiteId);
-            promise.success(function(status) {
-               if (status == 'OK') {
-                   $location.url("/user/" + vm.userId + "/website");
-               } else {
-                   vm.error = "An error occurred trying to delete the website";
-               }
-            });
+
+            var answer = confirm("Delete this website?");
+            if (answer) {
+                var promise = WebsiteService.deleteWebsite(vm.websiteId);
+                promise.success(function (status) {
+                    if (status == 'OK') {
+                        $location.url("/user/" + vm.userId + "/website");
+                    }
+                }).error(function (err) {
+                    vm.error = "An uncaught error occurred deleting your website: \n" + err;
+                });
+            }
         }
 
         function updateWebsite(newSite) {
             var promise = WebsiteService.updateWebsite(vm.websiteId, newSite);
             promise.success(function(status) {
-               if (status == 'OK') {
-                   $location.url("/user/" + vm.userId + "/website");
-               } else {
-                   vm.error = "An error occurred trying to update the website";
-               }
+                if (status == 'OK') {
+                    $location.url("/user/" + vm.userId + "/website");
+                }
+            }).error(function(err) {
+                if (err == 'Conflict') {
+                    vm.error = "A website with that name already exists! Please use a different name.";
+                } else {
+                    vm.error = "An uncaught error occurred when updating your website: \n" + err;
+                }
             });
         }
     }
